Made a use-zoom hook and removed duplicate code
